fix: fetch correct balance after network switch
